Mark a partially constructed failed DefaultMcpClient as closed (#2977)
This should prevent a resource leak that could occur if McpTransport
invokes the onFailure() callback. In such a case, DefaultMcpClient might
attempt to reinitialize itself, even though its constructor originally
threw an exception. Since the caller would have no reference to the
DefaultMcpClient instance, they would be unable to use or explicitly
close the client, leading to a resource leak.