Exposing MCP resources automatically over synthetic tools (#3302)
Fixes https://github.com/langchain4j/langchain4j/issues/2767

Wanted to get some early feedback before I proceed to creating more
tests and documentation (and probably extend it to prompts too).

The idea is that the user can now, when creating a `McpToolProvider`,
enable the `exposeResourcesAsTools` option. If this is enabled, then the
tool provider will expose two extra 'synthetic' tools:

- `list_resources` to obtain a list of all resources available for all
MCP clients that are wrapped into this tool provider
- `get_resource` to render a specific resource - it takes two arguments:
- resource URI (I chose URI instead of the name to be able to retrieve
[resource
templates](https://modelcontextprotocol.io/specification/2025-06-18/server/resources#resource-templates)
too in the same way)
  - name (`key`) of the MCP client that this resource belongs to

The result of `list_resources` is JSON and looks something like this:

```
[ {
  "mcpServer" : "alice",
  "uri" : "file:///info",
  "uriTemplate" : null,
  "name" : "basicInfo",
  "description" : "Basic information about Alice",
  "mimeType" : "text/plain"
}, {
  "mcpServer" : "bob",
  "uri" : "file:///info",
  "uriTemplate" : null,
  "name" : "basicInfo",
  "description" : "Basic information about Bob",
  "mimeType" : "text/plain"
} ]
```

I tested this with `gpt-4o-mini` and it understood how to find and
retrieve a relevant resource (see
`McpResourcesAsToolsTestBase#integrationTestWithAiService` for
demonstration)

So a resource is identified by uri AND the name of the mcp connection,
because it may happen that two servers will have resources on the same
URI.
When it's a resource, then `uri` is specified here, if it's a template,
then `uriTemplate` is specified instead.

Note: another way would be to expose a separate synthetic tool for each
resource instead of having `get_resource` over all of them. That would
render it unnecessary to invent a new JSON schema for the
`list_resources`, but I'm afraid that there will, in the real world, be
too many resources, and exposing a tool for each separately would
explode the context size.