Rework MCP client to support structured content results (#3733)
Fixes #3714 

Also bumps quarkus-mcp-server for integration tests to 1.5.3

@dliubars I'm not completely sure about changing the return type of
`McpClient.executeTool`, if you prefer I can instead add a new method
(`executeToolStructured`?) and keep the original one that returns
String. The problem is I don't like the name `executeToolStructured`
(better idea?) and it feels like it would unnecessarily clutter the API
a bit.

The best long-term strategy is probably to have just `executeTool` that
returns a `ToolExecutionResult`, but there's no way to achieve that
without a breaking change. I would say most people use a
`McpToolProvider` instead of directly working with `McpClient`, so the
breakage should be relatively minor.