Make GoogleAiGeminiChatModel unit-testable (#3927)
## Issue
Preparation for: #3916 

## Change
Makes the `GoogleAiGeminiChatModel` testable by allowing the
`GeminiService` to be passed in as ctor param. Other than that, I've
added some basic coverage to the `GeminiService` and
`GoogleAiGeminiChatModel` - a lot of defaults or edge-cases are not
tested in the unit tests.

Production code changes:
* Require the `GeminiService` to be passed into the Base client - I've
opted for this opposed to using the `ServiceLoader` for simplicity let
me know if you think `ServiceLoader` is a better option here
* Changed `GeminiGenerationConfig` to be a record so that it can easily
be compared in tests.


## General checklist
<!-- Please double-check the following points and mark them like this:
[X] -->
- [x] There are no breaking changes (API, behaviour)
- [x] I have added unit and/or integration tests for my change
- [x] The tests cover both positive and negative cases
- [x] I have manually run all the unit and integration tests in the
module I have added/changed, and they are all green
- [x] I have manually run all the unit and integration tests in the
[core]