Refactor value classes to (nested) records in google-ai-gemini (#4070)
## Change
This further flattens the classes inside the `google-ai-gemini` module.
I've started writing changes to implement the Embedding Batch API for
Gemini (https://ai.google.dev/gemini-api/docs/batch-api#batch-embedding)
but the preexisting batch embedding API makes it difficult to follow,
and it would be unclear which one is a async batch and which is
synchronous.

These changes are similar to the ones in #3955 but are around the non
chat model requests/responses. After these only (public) models are
prefixed with `GoogleAi` everything else is prefixed Gemini. So this
leaves the (currently) 4 model services:
 * `GoogleAiEmbeddingModel.java` - Synchronous batching
 * `GoogleAiGeminiChatModel.java` - Default chat model
 * `GoogleAiGeminiStreamingChatModel.java` - Streaming chat model
 * `GoogleAiGeminiTokenCountEstimator.java` - Token counter
 * `GoogleAiGeminiBatchChatModel` - Batching in #3942

## General checklist
- [x] There are no breaking changes (API, behaviour)
- [x] I have added unit and/or integration tests for my change
- [x] The tests cover both positive and negative cases
- [x] I have manually run all the unit and integration tests in the
module I have added/changed, and they are all green
- [x] I have manually run all the unit and integration tests in the
[core](https://github.com/langchain4j/langchain4j/tree/main/langchain4j-core)
and
[main](https://github.com/langchain4j/langchain4j/tree/main/langchain4j)
modules, and they are all green